+++ core/src/main/java/io/questdb/cairo/mig/EngineMigration.java
@@ -54,9 +54,9 @@ public class EngineMigration {
         long mem = Unsafe.malloc(tempMemSize, MemoryTag.NATIVE_MIG);
 
         try (
-                MemoryARW virtualMem = Vm.getARWInstance(ff.getPageSize(), Integer.MAX_VALUE, MemoryTag.NATIVE_MIG_MMAP);
-                Path path = new Path();
-                MemoryMARW rwMemory = Vm.getMARWInstance()
+            MemoryARW virtualMem = Vm.getARWInstance(ff.getPageSize(), Integer.MAX_VALUE, MemoryTag.NATIVE_MIG_MMAP);
+            Path path = new Path();
+            MemoryMARW rwMemory = Vm.getMARWInstance()
         ) {
             MigrationContext context = new MigrationContext(engine, mem, tempMemSize, virtualMem, rwMemory);
             path.of(configuration.getRoot());
@@ -64,18 +64,18 @@ public class EngineMigration {
             // check if all tables have been upgraded already
             path.concat(TableUtils.UPGRADE_FILE_NAME).$();
             final boolean existed = !force && ff.exists(path);
-            long upgradeFd = openFileRWOrFail(ff, path, configuration.getWriterFileOpenOpts());
+            int upgradeFd = openFileRWOrFail(ff, path, configuration.getWriterFileOpenOpts());
             LOG.debug()
-                    .$("open [fd=").$(upgradeFd)
-                    .$(", path=").$(path)
-                    .$(']').$();
+                .$("open [fd=").$(upgradeFd)
+                .$(", path=").$(path)
+                .$(']').$();
             if (existed) {
                 int currentVersion = TableUtils.readIntOrFail(
-                        ff,
-                        upgradeFd,
-                        0,
-                        mem,
-                        path
+                    ff,
+                    upgradeFd,
+                    0,
+                    mem,
+                    path
                 );
 
                 if (currentVersion >= latestVersion) {
@@ -90,20 +90,20 @@ public class EngineMigration {
                     LOG.info().$("upgrading database [version=").$(latestVersion).I$();
                     if (upgradeTables(context, latestVersion)) {
                         TableUtils.writeIntOrFail(
-                                ff,
-                                upgradeFd,
-                                0,
-                                latestVersion,
-                                mem,
-                                path
+                            ff,
+                            upgradeFd,
+                            0,
+                            latestVersion,
+                            mem,
+                            path
                         );
                     }
                 } finally {
                     Vm.bestEffortClose(
-                            ff,
-                            LOG,
-                            upgradeFd,
-                            Integer.BYTES
+                        ff,
+                        LOG,
+                        upgradeFd,
+                        Integer.BYTES
                     );
                 }
             }
@@ -137,15 +137,15 @@ public class EngineMigration {
                     path.concat(TableUtils.META_FILE_NAME);
 
                     if (ff.exists(path.$())) {
-                        final long fd = openFileRWOrFail(ff, path, context.getConfiguration().getWriterFileOpenOpts());
+                        final int fd = openFileRWOrFail(ff, path, context.getConfiguration().getWriterFileOpenOpts());
                         try {
                             int currentTableVersion = TableUtils.readIntOrFail(ff, fd, META_OFFSET_VERSION, mem, path);
                             if (currentTableVersion < latestVersion) {
                                 LOG.info()
-                                        .$("upgrading [path=").$(path)
-                                        .$(",fromVersion=").$(currentTableVersion)
-                                        .$(",toVersion=").$(latestVersion)
-                                        .I$();
+                                    .$("upgrading [path=").$(path)
+                                    .$(",fromVersion=").$(currentTableVersion)
+                                    .$(",toVersion=").$(latestVersion)
+                                    .I$();
 
                                 copyPath.trimTo(plen);
                                 backupFile(ff, path, copyPath, TableUtils.META_FILE_NAME, currentTableVersion);
@@ -162,20 +162,20 @@ public class EngineMigration {
                                             path.trimTo(plen);
                                         } catch (Throwable e) {
                                             LOG.error().$("failed to upgrade table path=")
-                                                    .$(path.trimTo(plen))
-                                                    .$(", exception: ")
-                                                    .$(e).$();
+                                                .$(path.trimTo(plen))
+                                                .$(", exception: ")
+                                                .$(e).$();
                                             throw e;
                                         }
                                     }
 
                                     TableUtils.writeIntOrFail(
-                                            ff,
-                                            fd,
-                                            META_OFFSET_VERSION,
-                                            ver,
-                                            mem,
-                                            path.trimTo(plen)
+                                        ff,
+                                        fd,
+                                        META_OFFSET_VERSION,
+                                        ver,
+                                        mem,
+                                        path.trimTo(plen)
                                     );
                                 }
                             }
